requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
feedparser>=6.0.11
beautifulsoup4>=4.12.3
matplotlib>=3.9.0
pandas>=2.2.2